+++ src/main/java/alignpro/Controller/ProjectOverview/SubTaskController.java
@@ -2,6 +2,7 @@ package alignpro.Controller.ProjectOverview;
 
 import alignpro.Model.Projects.SubTask;
 import alignpro.Service.SubTaskService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +18,8 @@ public class SubTaskController {
     }
 
     @GetMapping("/createSubTask/{taskID}")
-    public String createSubTask(@PathVariable("taskID") int taskID, Model model) {
+    public String createSubTask(@PathVariable("taskID") int taskID, Model model, HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         SubTask obj = new SubTask();
         obj.setTaskID(taskID);
         model.addAttribute("obj", obj);
@@ -25,13 +27,15 @@ public class SubTaskController {
     }
 
     @PostMapping("/saveSubTask")
-    public String saveSubTask(@ModelAttribute SubTask newSubTask){
+    public String saveSubTask(@ModelAttribute SubTask newSubTask, HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         subTaskService.saveSubTask(newSubTask);
         return "redirect:/";
     }
 
     @GetMapping("/edit-subTask/{subTaskID}")
-    public String editSubTask(@PathVariable("subTaskID") int subTaskID, Model model){
+    public String editSubTask(@PathVariable("subTaskID") int subTaskID, Model model, HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         SubTask obj = subTaskService.getSubTask(subTaskID);
         if(obj != null){
             model.addAttribute("obj", obj);
@@ -48,16 +52,23 @@ public class SubTaskController {
                                 @RequestParam("startDate") String startDate,
                                 @RequestParam("endDate") String endDate,
                                 @RequestParam("time") int time,
-                                @RequestParam("skillRequirement") String skillRequirement){
+                                @RequestParam("skillRequirement") String skillRequirement,
+                                HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         SubTask subTask = new SubTask(subTaskID, subTaskName, subTaskDescription, startDate, endDate, time, skillRequirement);
         subTaskService.editSubTask(subTask, subTask.getSubTaskID());
         return "redirect:/";
     }
 
     @PostMapping("/deleteSubTask/{subTaskID}")
-    public String deleteSubTask(@PathVariable("subTaskID") int subTaskID){
+    public String deleteSubTask(@PathVariable("subTaskID") int subTaskID, HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         subTaskService.deleteSubTask(subTaskID);
         return "redirect:/";
+    }
+
+    public boolean isUserLoggedIn(HttpSession session){
+        return session.getAttribute("pmUserID") == null;
     }
